Add help text

+++ GMailDeleteOlderThan/gmail-delete-older-than.js
@@ -1,3 +1,20 @@
+/*
+Script that lets you use GoogleScript engine to delete emails older than x days in particular labels.
+
+1. Go to script.google.com.
+2. Create a new project.
+3. Paste this script.
+6. Run. 
+7. View Logger - verify working as expected.
+8. Set debug to false.
+
+
+*/
+
+var debug = true; // true/false -> true will only log to console, false will actually delete.
+
+// These represent labels - for example the ".delete_7" label I want to delete emails older than 7 days.
+// You can add as many such labels as you like. I've commented the last two as an example 
 var filters = [
    ['.delete_7', 7]
   ,['.delete_30', 30]
@@ -7,7 +24,7 @@ var filters = [
 ];
 
 
-function purgeEmails(label, days) {
+function purgeEmails(label, days, debug) {
   var cutoff = new Date();
   cutoff.setDate(cutoff.getDate() - days)
 
@@ -23,9 +40,13 @@ function purgeEmails(label, days) {
       var messages = GmailApp.getMessagesForThread(threads[thread]);
       for (var message = 0; message < messages.length; message++) {
         var email = messages[message];       
-        if (email.getDate() < cutoff) {
-          // console.log(email.getFrom() + ": " + email.getSubject());
-          email.moveToTrash();
+        if (email.getDate() < cutoff) {          
+          if (debug) {
+            console.log(email.getFrom() + ": " + email.getSubject());
+          }
+          else {
+            email.moveToTrash();
+          }
           count++;
         }
       }
@@ -39,7 +60,7 @@ function purgeEmails(label, days) {
 
 function clearEmails() {
   console.log("Start")
-  for (var i = 0; i < filters.length; i++) {
-    purgeEmails(filters[i][0], filters[i][1]);
+  for (var filter = 0; filter < filters.length; filter++) {
+    purgeEmails(filters[filter][0], filters[filter][1], debug);
   }
 }
